backend/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
pymupdf>=1.24.0
faiss-cpu>=1.7.0
tqdm>=4.66.0
python-dotenv>=1.0.0
groq>=0.11.0

fastapi>=0.128.0
uvicorn>=0.40.0
sqlalchemy>=2.0.40
pydantic>=2.12.0
psycopg2-binary>=2.9.0
google-generativeai>=0.6.0
